+++ venndb-usage/tests/fails/any_bool.rs
@@ -1,0 +1,9 @@
+use venndb::VennDB;
+
+#[derive(Debug, VennDB)]
+struct Employee {
+    #[venndb(any)]
+    is_alive: bool,
+}
+
+fn main() {}