+++ venndb-usage/tests/fails/filter_any_bool.rs
@@ -1,0 +1,9 @@
+use venndb::VennDB;
+
+#[derive(Debug, VennDB)]
+struct Employee {
+    #[venndb(filter, any)]
+    is_alive: bool,
+}
+
+fn main() {}