+++ venndb-usage/tests/fails/key_any.rs
@@ -1,0 +1,9 @@
+use venndb::VennDB;
+
+#[derive(Debug, VennDB)]
+struct Employee {
+    #[venndb(key, any)]
+    country: String,
+}
+
+fn main() {}