+++ CHANGELOG.md
@@ -5,7 +5,58 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# 0.2.0 (2024-04-12)
+# 0.2.1 (2024-04-15)
+
+A backwards compatible patch for [v0.2.0](#020-2024-04-15),
+to support rows that allow any value for a specific column.
+
+Non-Breaking changes:
+
+* support `#[venndb(any)]` filters;
+  * these are possible only for `T` filter maps, where `T: ::venndb::Any`;
+  * `bool` filters cannot be `any` as `bool` doesn't implement the `::venndb::Any` trait;
+  * rows that are `any` will match regardless of the query filter used for that property;
+
+Example usage:
+
+```rust
+use venndb::{Any, VennDB};
+#[derive(Debug, PartialEq, Eq, Clone, Hash)]
+pub enum Department {
+  Any,
+  Hr,
+  Engineering,
+}
+
+impl Any for Department {
+  fn is_any(&self) -> bool {
+    self == Department::Any
+  }
+}
+
+#[derive(Debug, VennDB)]
+pub struct Employee {
+  name: String,
+  #[venndb(filter, any)]
+  department: Department,
+}
+
+let db = EmployeeDB::from_iter([
+  Employee { name: "Jack".to_owned(), department: Department::Any },
+  Employee { name: "Derby".to_owned(), department: Department::Hr },
+]);
+let mut query = db.query();
+
+// will match Jack and Derby, as Jack is marked as Any, meaning it can work for w/e value
+let hr_employees: Vec<_> = query.department(Department::Hr).execute().unwrap().iter().collect();
+assert_eq!(hr_employees.len(), 2);
+```
+
+In case you combine it with the filter map property being optional (`department: Option<Department>`),
+then it will still work the same, where rows with `None` are seen as nothing at all and just ignored.
+This has no affect on the correct functioning of `Any`.
+
+# 0.2.0 (2024-04-15)
 
 Breaking Changes:
 
@@ -22,9 +73,50 @@ While this changes behaviour of `filters` and `filter maps` it is unlikely that 
 `Option<T>` for these types before, as their ergonomics have been a bit weird prior to this version.
 Even more so for `filter maps` it could have resulted in panics.
 
+Options, be it filters of filter maps, allow you to have rows that do not register any value for optional
+properties, allowing them to exist without affecting the rows which do have it.
+
 Non-Breaking Changes:
 
 * improve documentation;
+
+Updated Example from 0.1:
+
+```rust
+use venndb::VennDB
+
+#[derive(Debug, VennDB)]
+pub struct Employee {
+    #[venndb(key)]
+    id: u32,
+    name: String,
+    is_manager: Option<bool>,
+    is_admin: bool,
+    #[venndb(skip)]
+    foo: bool,
+    #[venndb(filter)]
+    department: Department,
+    #[venndb(filter)]
+    country: Option<String>,
+}
+
+fn main() {
+    let db = EmployeeDB::from_iter(/* .. */);
+
+    let mut query = db.query();
+    let employee = query
+        .is_admin(true)
+        .is_manager(false)  // rows which have `None` for this property will NOT match this filter
+        .department(Department::Engineering)
+        .execute()
+        .expect("to have found at least one")
+        .any();
+
+    println!("non-manager admin engineer: {:?}", employee);
+    // as we didn't specify a `country` filter, even rows without a country specified will
+    // match here if they match the defined (query) filters)
+}
+```
 
 # 0.1.1 (2024-04-10)